+++ python/astra/training/__init__.py
@@ -1,18 +1,31 @@
 from astra.training.checkpoint import load_checkpoint, save_checkpoint
 from astra.training.data import Corpus, SeqStream, tokenize_corpus
-from astra.training.loop import TrainReport, train, val_loss
-from astra.training.optim import AdamW, CosineSchedule, clip_grad_norm
+from astra.training.loop import TrainReport, loss_by_shard, train, val_loss
+from astra.training.optim import (
+    OPTIMIZER_REGISTRY,
+    SCHEDULE_REGISTRY,
+    AdamW,
+    CosineSchedule,
+    build_optimizer,
+    build_schedule,
+    clip_grad_norm,
+)
 
 __all__ = [
+    "OPTIMIZER_REGISTRY",
+    "SCHEDULE_REGISTRY",
     "AdamW",
     "Corpus",
     "CosineSchedule",
     "SeqStream",
     "TrainReport",
+    "build_optimizer",
+    "build_schedule",
+    "clip_grad_norm",
+    "load_checkpoint",
+    "loss_by_shard",
+    "save_checkpoint",
     "tokenize_corpus",
     "train",
     "val_loss",
-    "clip_grad_norm",
-    "load_checkpoint",
-    "save_checkpoint",
 ]